pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "signature-packet"
version = "0.1.0"
description = "Extract signature pages from DOCX/PDF using Tesseract OCR and merge into one PDF"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "PyMuPDF>=1.23.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
gui = ["tkinterdnd2>=0.4.0"]

[project.scripts]
signature-packet = "signature_packet.cli:main"
signature-packet-gui = "signature_packet.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["signature_packet*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
